Add files via upload
actuator_control uses UDP to receive the packets and control the motors. there are two arrays, LPWM_list and RPWM_list. in those, each element is the pin number for the LPWM and RPWM signals. note that the elements have to match. so if LPWM of a motor driver is in LPWM[2], then RPWM[2] should have the pin number of the RPWM for the same motor driver. they cant be randomly added in the array. one possible error that might occur is that the function "stop_all_motors()" might not work in some sections of the code during operation. in which case, replace that function call in that line where its not working with a for loop that analogwrites 0 to all the pins.